test(struct-init-v2): repartition return-effect coverage
