+++ testdata/src/go.uber.org/structinitv2/limitations/limitations.go
@@ -172,38 +172,6 @@ func tForwardViaFuncValue() *int {
 	return b.child.ptr
 }
 
-// False negative: a struct-returning call forwarded through a return (`return f()`) is unsupported.
-// The callee's per-field return summary is not composed across the forward, so the nil child from
-// makeNilNode is not tracked at the caller. Supporting it would need the return-read demand closed
-// over forwarding edges (as closeParamFieldSets does for params), which is not computed for returns.
-func makeNilNode() *Node { return &Node{} }
-
-func forwardNilNode() *Node { return makeNilNode() }
-
-func tForwardedReturnNotComposed() *int {
-	b := forwardNilNode()
-	return b.child.ptr
-}
-
-// False negative: a struct-returning call used directly as an argument is not composed into the
-// parameter boundary. Return-effects support must bind makeNilNodeForParam's field summary to n.
-func makeNilNodeForParam() *Node { return &Node{} }
-
-func readReturnedParam(n *Node) *int { return n.child.ptr }
-
-func tReturnedCallAsParam() *int {
-	return readReturnedParam(makeNilNodeForParam())
-}
-
-// False negative: the same missing return-to-boundary composition applies to a method receiver.
-func makeNilNodeForReceiver() *Node { return &Node{} }
-
-func (n *Node) readReturnedReceiver() *int { return n.child.ptr }
-
-func tReturnedCallAsReceiver() *int {
-	return makeNilNodeForReceiver().readReturnedReceiver()
-}
-
 // ---------------------------------------------------------------------------------------------
 // FALSE POSITIVES (over-reports — //want required)
 // ---------------------------------------------------------------------------------------------
