+++ testdata/src/go.uber.org/structinitv2/returnlocal/functionreturntransitive.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package funcreturnfields
+package returnlocal
 
 // Direct return of a composite literal local.
