+++ testdata/src/go.uber.org/structinitv2/returnlocal/methodreturnbasic.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package funcreturnfields
+package returnlocal
 
 // Same as funcreturnbasic.go, but for methods.
